Clarify prepared SVD lint invariants
Why: performance-motivated lint exceptions need adjacent rationale, while unit tests belong in the repository's dedicated module layout for focused maintenance.
